First commit!
It will connect to a server, send the initial messages (incl. JOIN),
and echo what it receives from the server. And that’s it.

Next step would be to handle the PING command. Also command line args
or a config file (to gitignore) for all these, rather than hardcoded,
esp for PASS.